pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meshcore-tcp-bot"
version = "0.1.0"
description = "MeshCore TCP bot for XIAO WiFi RS232Bridge repeaters"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "cryptography>=45.0.0,<46.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "itsdangerous>=2.2.0,<3.0.0",
  "pynacl>=1.5.0,<2.0.0",
  "uvicorn>=0.30.0,<1.0.0",
]

[project.scripts]
meshcore-tcp-bot = "meshcore_tcp_bot.__main__:main"

[tool.setuptools.packages.find]
include = ["meshcore_tcp_bot*"]